+++ deltafi-ui/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Bug where `Total File Size` field on DeltaFile viewer was showing wrong value (#199)
+
+## [0.96.0]
+
 ### Changed
 
 - Added support for Basic Authentication on Users page (#191)
@@ -276,7 +282,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - NiFi link from Dashboard
 - Requested resource information from System Metrics
 
-[Unreleased]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.95.4...main
+[Unreleased]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.96.0...main
+[0.96.0]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.95.4...0.96.0
 [0.95.4]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.95.3...0.95.4
 [0.95.3]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.95.2...0.95.3
 [0.95.2]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.95.1...0.95.2